Move daemonchecker component
