+++ .github/CODEOWNERS
@@ -342,8 +342,8 @@
 /comp/metadata/packagesigning @DataDog/agent-delivery
 /comp/trace/etwtracer @DataDog/windows-agent
 /comp/updater/ssistatus @DataDog/fleet
+/comp/updater/daemonchecker @DataDog/fleet
 /comp/autoscaling/datadogclient @DataDog/container-integrations
-/comp/daemonchecker @DataDog/fleet
 /comp/etw @DataDog/windows-agent
 /comp/fleetstatus @DataDog/fleet
 /comp/haagent @DataDog/ndm-core
