+++ comp/updater/daemonchecker/def/component.go
