+++ comp/updater/daemonchecker/impl/daemon_checker_windows.go
